+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    were completely changed to reflect the real state of things.
 2. `Package::targetings` is now of type `CampaignTargeting` rather than `Targeting`.
 3. `Dsl\MyTarget\Domain\V1\Targeting\Targeting` class was removed.
+4. `RemarketingItem::usersLists` type was changed to `RemarketingItemUserList`
 
 #### Fixes